rename more modules, remove public module
