+++ src/business_logic/transaction/objects/internal_invoke_function.rs
@@ -17,10 +17,10 @@ use crate::{
         calculate_transaction_hash_common, TransactionHashPrefix,
     },
     definitions::{
-        constants::EXECUTE_ENTRY_POINT_SELECTOR, general_config::StarknetGeneralConfig,
+        constants::{EXECUTE_ENTRY_POINT_SELECTOR, VALIDATE_ENTRY_POINT_SELECTOR},
+        general_config::StarknetGeneralConfig,
         transaction_type::TransactionType,
     },
-    public::abi::VALIDATE_ENTRY_POINT_SELECTOR,
     services::api::contract_classes::deprecated_contract_class::EntryPointType,
     utils::{calculate_tx_resources, Address},
 };
