+++ src/definitions/constants.rs
@@ -80,4 +80,7 @@ lazy_static! {
     /// Value generated from `get_selector_from_name('Transfer')`.
     pub static ref TRANSFER_EVENT_SELECTOR: Felt252 =
         felt_str!("271746229759260285552388728919865295615886751538523744128730118297934206697");
+
+    pub static ref VALIDATE_ENTRY_POINT_SELECTOR: Felt252 =
+        felt_str!("626969833899987279399947180575486623810258720106406659648356883742278317941");
 }
