+++ src/lib.rs
@@ -11,10 +11,9 @@ pub mod core;
 pub mod definitions;
 pub mod hash_utils;
 pub mod parser_errors;
-pub mod public;
+pub mod runner;
 pub mod services;
-pub mod starknet_runner;
-pub mod starknet_storage;
 pub mod starkware_utils;
+pub mod storage;
 pub mod testing;
 pub mod utils;
