+++ src/services/api/contract_classes/deprecated_contract_class.rs
@@ -1,4 +1,4 @@
-use crate::{public::abi::AbiType, services::api::contract_class_errors::ContractClassError};
+use crate::services::api::contract_class_errors::ContractClassError;
 use cairo_vm::felt::{Felt252, PRIME_STR};
 use cairo_vm::{
     serde::deserialize_program::{
@@ -12,6 +12,7 @@ use cairo_vm::{
 };
 use getset::{CopyGetters, Getters};
 use starknet_api::deprecated_contract_class::EntryPoint;
+use starknet_contract_class::AbiType;
 use std::{collections::HashMap, fs::File, io::BufReader, path::PathBuf};
 
 pub(crate) const SUPPORTED_BUILTINS: [BuiltinName; 5] = [
